feat(content): adding authorId to new's comments

+++ content/src/services/comments.service.ts
@@ -1,19 +1,33 @@
 import { prisma } from "../lib/pisma.js"
 
 
-export const addCommentService = async (newId: string, comment: string) => {
+export const addCommentService = async (newId: string, comment: string, userId: string) => {
     return await prisma.comments.create({
         data: {
             content: comment,
-            newsId: newId
+            newsId: newId,
+            authorId: userId
         }
     })
 }
 
-export const updateCommentService = async (commentId: string, newContent: string) => {
+export const updateCommentService = async (commentId: string, newContent: string, userId: string) => {
+    const comment = await prisma.comments.findUnique({
+        where: {
+            id: commentId
+        }
+    });
+
+    if (!comment) {
+        throw new Error("Commentaire introuvable")
+    }
+
+    if (comment.authorId !== userId) {
+        throw new Error("Non autorisé, vous ne pouvez pas mettre ce commentaire à jour")
+    }
     return await prisma.comments.update({
         where: { 
-            id: commentId 
+            id: commentId
         },
         data: {
             content: newContent
@@ -21,10 +35,24 @@ export const updateCommentService = async (commentId: string, newContent: string
     })
 }
 
-export const deleteCommentService = async (commentId: string) => {
+export const deleteCommentService = async (commentId: string, userId: string) => {
+    const comment = await prisma.comments.findUnique({
+        where: {
+            id: commentId
+        }
+    });
+
+    if (!comment) {
+        throw new Error("Commentaire introuvable")
+    }
+
+    if (comment?.authorId !== userId) {
+        throw new Error("Non autorisé, vous ne pouvez pas supprimer ce commentaire");
+    }
     return await prisma.comments.delete({
         where: { 
-            id: commentId
+            id: commentId,
+            authorId: userId
          }
     })
 }
